requirements.txt
anthropic>=0.49.0
python-dotenv>=1.1.0
pydantic>=2.11.1
pydantic-settings>=2.8.1
setuptools>=69.2.0

# Development dependencies
pylint>=3.0.3
black>=24.2.0
